Final Update
If you see this sorry for the lack of commits, I am not great with GitHub and am working on it but I struggle with keeping track of updating so I haven't. I just read a bit ago when finalizing my project that 3 commits over time was a requirement and I hope you can understand. If not I take accountability, I should've read closer earlier.